fix: backend production readiness fixes
- core/config.py: add python-dotenv, safe defaults for DATABASE_URL/REDIS_URL, JWT_SECRET raises in prod
- routes/auth.py: add org_id to SignupResponse, fix timezone-naive datetime comparison
- routes/billing.py: fix float truncation, balance_credits/credits_added as float
- routes/agents.py: add validation_alias so model_config serializes correctly from ORM
- routes/calls.py: add charge_final_credits for short call billing, extract to helper
- routes/monitor.py: add 30s ping/timeout to prevent goroutine leak on silent disconnect
- routes/telephony.py: fix inbound_telnyx to return proper XML response, add saxutils XML escaping, fix Twilio WS handler to use correct pipeline API
- services/dialer_service.py: replace hardcoded caller ID with PhoneNumber DB lookup + env fallback, use ENVIRONMENT flag for simulation
- services/monitor_tracker.py: replace deprecated get_event_loop() with get_running_loop()
- services/transfer_manager.py: replace getattr agent.phone_number with PhoneNumber table lookup, fix get_event_loop() deprecation
- services/voicemail_detection.py: add proper WAV header to PCM before sending to Whisper
- services/pipeline/transport/webrtc_transport.py: minor cleanup
- services/rag_service.py: minor cleanup
- tests/conftest.py: add disable_rate_limit autouse fixture, override_auth fixture, import get_current_user/org for dependency_overrides
- tests/test_auth.py: update to match new SignupResponse with org_id
- tests/test_rag_and_campaigns.py: use override_auth fixture instead of module-level patch
- tests/test_telephony.py: update test_inbound_telnyx to expect XML response
- tests/test_transfer_manager.py: add third mock for PhoneNumber query
- requirements.txt: remove duplicate twilio/audioop-lts, add aiosqlite
- docker-compose.yml: switch to pgvector/pgvector:pg16, add ARQ worker service, add coturn, remove --reload
- .env.example: add TELNYX_CONNECTION_ID, TELNYX_CALLER_ID entries

All 42 tests passing.

+++ backend/app/routes/billing.py
@@ -32,8 +32,8 @@ class VerifyPaymentRequest(BaseModel):
 
 class VerifyPaymentResponse(BaseModel):
     success: bool
-    credits_added: int
-    new_balance: int
+    credits_added: float
+    new_balance: float
 
 class CreditTransactionResponse(BaseModel):
     id: int
@@ -50,7 +50,7 @@ class CreditTransactionResponse(BaseModel):
         from_attributes = True
 
 class BalanceResponse(BaseModel):
-    balance_credits: int
+    balance_credits: float
     transactions: list[CreditTransactionResponse]
 
 # ---------- Helper ----------
@@ -117,11 +117,11 @@ async def verify_payment(
     ct.status = "completed"
     ct.completed_at = datetime.now(UTC)
     ct.razorpay_payment_id = payload.razorpay_payment_id
-    credits = int(ct.credits_added)
-    org.balance_credits = (org.balance_credits or 0) + credits
+    credits = float(ct.credits_added)
+    org.balance_credits = (org.balance_credits or 0.0) + credits
     db.add(org)  # ensure org update is tracked
     await db.commit()
-    return VerifyPaymentResponse(success=True, credits_added=credits, new_balance=int(org.balance_credits))
+    return VerifyPaymentResponse(success=True, credits_added=credits, new_balance=float(org.balance_credits))
 
 @router.post("/billing/razorpay/webhook")
 async def razorpay_webhook(
@@ -163,7 +163,7 @@ async def razorpay_webhook(
             )
             org = org_result.scalar_one_or_none()
             if org:
-                org.balance_credits = (org.balance_credits or 0) + credits
+                org.balance_credits = (org.balance_credits or 0.0) + credits
                 db.add(org)
             await db.commit()
             return {"status": "processed", "credits_added": credits}
@@ -182,5 +182,5 @@ async def get_balance(
         .order_by(CreditTransaction.created_at.desc())
     )
     transactions = result.scalars().all()
-    return BalanceResponse(balance_credits=int(org.balance_credits or 0), transactions=transactions)
+    return BalanceResponse(balance_credits=float(org.balance_credits or 0.0), transactions=transactions)
